Optimize the FileUtils

+++ src/com/jamiexu/utils/file/FileUtils.java
@@ -1,18 +1,23 @@
 package com.jamiexu.utils.file;
 
+import com.jamiexu.utils.convert.ByteUtils;
 import com.jamiexu.utils.convert.ConvertUtils;
 
 import java.io.*;
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
 import java.util.Base64;
+import java.util.Objects;
 import java.util.zip.CRC32;
 
 /**
- * @author Jamiexu or Jamie793
- * @version 1.0
- * 博客 blog.jamiexu.cn
- */
+ * @author Jamiexu/Jamie793
+ * @version 3.0
+ * @date 2020/9/4
+ * @time 19:11
+ * @blog https://blog.jamiexu.cn
+ **/
+
 
 public class FileUtils {
 
@@ -55,7 +60,7 @@ public class FileUtils {
             }
 
         }
-        return stringBuilder.toString();
+        return stringBuilder.length() > 0 ? stringBuilder.toString() : null;
     }
 
 
@@ -101,17 +106,19 @@ public class FileUtils {
             }
 
         }
-        return stringBuilder.toString();
+        return stringBuilder.length() > 0 ? stringBuilder.toString() : null;
     }
+
 
     /**
      * 获取指定行的文本内容
+     *
      * @param path 文件路径
      * @param line 行号
      * @return
      */
-    public static String getLineString(String path,int line){
-        return getLinesString(path,line,line);
+    public static String getLineString(String path, int line) {
+        return getLinesString(path, line, line);
     }
 
 
@@ -120,14 +127,15 @@ public class FileUtils {
      *
      * @param path    需要写出的文件路径
      * @param content 需要写出的文件内容
+     * @return boolean 状态
      */
-    public static void putString(String path, String content) {
+    public static boolean putString(String path, String content) {
+        boolean status = false;
         FileWriter fileWriter = null;
-        BufferedWriter bufferedWriter = null;
         try {
             fileWriter = new FileWriter(path);
-            bufferedWriter = new BufferedWriter(fileWriter);
-            bufferedWriter.write(content);
+            fileWriter.write(content);
+            status = true;
         } catch (IOException e) {
             e.printStackTrace();
         } finally {
@@ -140,15 +148,8 @@ public class FileUtils {
                 }
             }
 
-            if (bufferedWriter != null) {
-                try {
-                    bufferedWriter.flush();
-                    bufferedWriter.close();
-                } catch (IOException e) {
-                    e.printStackTrace();
-                }
-            }
         }
+        return status;
     }
 
     /**
@@ -156,27 +157,31 @@ public class FileUtils {
      *
      * @param path  需要写出文件的路径
      * @param bytes 需要写出的内容
+     * @return boolean 状态
      */
-    public static void writeFile(String path, byte[] bytes) {
-        DataOutputStream dataOutputStream = null;
+    public static boolean writeFile(String path, byte[] bytes) {
+        boolean status = false;
+        FileOutputStream fileOutputStream = null;
         try {
-            dataOutputStream = new DataOutputStream(new FileOutputStream(path));
-            dataOutputStream.write(bytes);
-        } catch (FileNotFoundException e) {
-            e.printStackTrace();
+            fileOutputStream = new FileOutputStream(path);
+            fileOutputStream.write(bytes);
+            if (new File(path).exists())
+                status = true;
         } catch (IOException e) {
             e.printStackTrace();
         } finally {
-            if (dataOutputStream != null) {
+            if (fileOutputStream != null) {
                 try {
-                    dataOutputStream.flush();
-                    dataOutputStream.close();
+                    fileOutputStream.flush();
+                    fileOutputStream.close();
                 } catch (IOException e) {
                     e.printStackTrace();
                 }
             }
         }
+        return status;
     }
+
 
     /**
      * 读取二进制文件
@@ -185,20 +190,32 @@ public class FileUtils {
      * @return byte[] 文件内容
      */
     public static byte[] readFile(String path) {
-        DataInputStream dataInputStream = null;
+        FileInputStream fileInputStream = null;
+        ByteArrayOutputStream byteArrayOutputStream = null;
         try {
-            dataInputStream = new DataInputStream(new FileInputStream(path));
-            byte[] bytes = new byte[dataInputStream.available()];
-            dataInputStream.read(bytes);
-            return bytes;
-        } catch (FileNotFoundException e) {
-            e.printStackTrace();
+            fileInputStream = new FileInputStream(path);
+            byteArrayOutputStream = new ByteArrayOutputStream();
+            byte[] byt = new byte[1024 * 8];
+            int len;
+            while ((len = fileInputStream.read(byt)) != -1) {
+                byteArrayOutputStream.write(byt, 0, len);
+            }
+            return byteArrayOutputStream.toByteArray();
         } catch (IOException e) {
             e.printStackTrace();
         } finally {
-            if (dataInputStream != null) {
+            if (fileInputStream != null) {
                 try {
-                    dataInputStream.close();
+                    fileInputStream.close();
+                } catch (IOException e) {
+                    e.printStackTrace();
+                }
+            }
+
+            if (byteArrayOutputStream != null) {
+                try {
+                    byteArrayOutputStream.flush();
+                    byteArrayOutputStream.close();
                 } catch (IOException e) {
                     e.printStackTrace();
                 }
@@ -212,8 +229,10 @@ public class FileUtils {
      *
      * @param from 源文件路径
      * @param to   目标文件路径
+     * @return boolean 状态
      */
-    public static void copyFile(String from, String to) {
+    public static boolean copyFile(String from, String to) {
+        boolean status = false;
         FileInputStream fileInputStream = null;
         FileOutputStream fileOutputStream = null;
         try {
@@ -224,8 +243,9 @@ public class FileUtils {
             while ((len = fileInputStream.read(bytes, 0, bytes.length)) != -1) {
                 fileOutputStream.write(bytes, 0, len);
             }
-        } catch (FileNotFoundException e) {
-            e.printStackTrace();
+            if (new File(to).exists())
+                if (Objects.equals(FileUtils.getFileMd5(from, false), FileUtils.getFileMd5(to, false)))
+                    status = true;
         } catch (IOException e) {
             e.printStackTrace();
         } finally {
@@ -246,24 +266,29 @@ public class FileUtils {
                 }
             }
         }
+        return status;
     }
 
     /**
      * 删除文件目录
      *
      * @param path 需要删除的文件目录路径
+     * @return boolean
      */
-    public static void removeDirectory(String path) {
-        File[] files = new File(path).listFiles();
-        for (File f : files) {
-            if (f.isFile())
-                f.delete();
-            else {
+    public static boolean removeDirectory(String path) {
+        File file = new File(path);
+        if (!file.exists() || file.isFile())
+            return false;
+        File[] files = file.listFiles();
+        for (File f : Objects.requireNonNull(files)) {
+            if (f.isFile()) {
+                if (!f.delete())
+                    return false;
+            } else {
                 removeDirectory(f.toString());
-                f.delete();
             }
         }
-        new File(path).delete();
+        return new File(path).delete();
     }
 
     /**
@@ -290,6 +315,31 @@ public class FileUtils {
         return null;
     }
 
+
+    /**
+     * 获取文件MD5默认结果为小写
+     *
+     * @param path 文件路径
+     * @return String 16进制文本
+     */
+    public static String getFileMd5(String path) {
+        FileInputStream fileInputStream = null;
+        try {
+            fileInputStream = new FileInputStream(path);
+            MessageDigest messageDigest = MessageDigest.getInstance("MD5");
+            byte[] bytes = new byte[1024 * 8];
+            int len = -1;
+            while ((len = fileInputStream.read(bytes, 0, bytes.length)) != -1) {
+                messageDigest.update(bytes, 0, len);
+            }
+            return ConvertUtils.bytesToHex(messageDigest.digest(), false);
+        } catch (NoSuchAlgorithmException | IOException e) {
+            e.printStackTrace();
+        }
+        return null;
+    }
+
+
     /**
      * 获取文件Sha1
      *
@@ -313,6 +363,31 @@ public class FileUtils {
         }
         return null;
     }
+
+
+    /**
+     * 获取文件Sha1默认结果为小写
+     *
+     * @param path 文件路径
+     * @return String 16进制文本
+     */
+    public static String getFileSha1(String path) {
+        FileInputStream fileInputStream = null;
+        try {
+            fileInputStream = new FileInputStream(path);
+            MessageDigest messageDigest = MessageDigest.getInstance("SHA-1");
+            byte[] bytes = new byte[1024 * 8];
+            int len = -1;
+            while ((len = fileInputStream.read(bytes, 0, bytes.length)) != -1) {
+                messageDigest.update(bytes, 0, len);
+            }
+            return ConvertUtils.bytesToHex(messageDigest.digest(), false);
+        } catch (NoSuchAlgorithmException | IOException e) {
+            e.printStackTrace();
+        }
+        return null;
+    }
+
 
     /**
      * 获取文件Sha256
@@ -338,13 +413,38 @@ public class FileUtils {
         return null;
     }
 
+
+    /**
+     * 获取文件Sha256默认结果为小写
+     *
+     * @param path 文件路径
+     * @return String 16进制文本
+     */
+    public static String getFileSha256(String path) {
+        FileInputStream fileInputStream = null;
+        try {
+            fileInputStream = new FileInputStream(path);
+            MessageDigest messageDigest = MessageDigest.getInstance("SHA-256");
+            byte[] bytes = new byte[1024 * 8];
+            int len = -1;
+            while ((len = fileInputStream.read(bytes, 0, bytes.length)) != -1) {
+                messageDigest.update(bytes, 0, len);
+            }
+            return ConvertUtils.bytesToHex(messageDigest.digest(), false);
+        } catch (NoSuchAlgorithmException | IOException e) {
+            e.printStackTrace();
+        }
+        return null;
+    }
+
+
     /**
      * 获取文件CRC32
      *
      * @param path 文件路径
      * @return Long
      */
-    public static Long getFileCrc32(String path) {
+    public static long getFileCrc32(String path) {
         FileInputStream fileInputStream = null;
         try {
             fileInputStream = new FileInputStream(path);
@@ -358,16 +458,64 @@ public class FileUtils {
         } catch (IOException e) {
             e.printStackTrace();
         }
-        return null;
+        return -1;
     }
 
+
     /**
-     * 获取文件Base64
+     * 获取文件Base64,并把编码数据写出
      *
-     * @param path 文件路径
-     * @return String Base64文本
+     * @param path   文件路径
+     * @param target 写出编码后数据文件路径
+     * @return boolean 状态
      */
-    public static String getFileBase64(String path) {
+    public static boolean getFileBase64AndWrite(String path, String target) {
+        boolean status = false;
+        Base64.Encoder encoder = Base64.getEncoder();
+        FileInputStream fileInputStream = null;
+        PrintWriter printWriter = null;
+        try {
+            fileInputStream = new FileInputStream(path);
+            printWriter = new PrintWriter(new FileOutputStream(target));
+            byte[] byts = new byte[1024 * 8];
+            int len;
+            while ((len = fileInputStream.read(byts)) != -1) {
+                if (len != byts.length) {
+                    byte[] copy = ByteUtils.copyBytes(byts, 0, len);
+                    printWriter.println(encoder.encodeToString(copy));
+                } else {
+                    printWriter.println(encoder.encodeToString(byts));
+                }
+            }
+            if (new File(target).exists())
+                status = true;
+        } catch (IOException e) {
+            e.printStackTrace();
+        } finally {
+            if (fileInputStream != null) {
+                try {
+                    fileInputStream.close();
+                } catch (IOException e) {
+                    e.printStackTrace();
+                }
+            }
+
+            if (printWriter != null) {
+                printWriter.flush();
+                printWriter.close();
+            }
+        }
+        return status;
+    }
+
+
+    /**
+     * 获取小文件的Base64并返回结果
+     *
+     * @param path 待获取文件路径
+     * @return String 结果
+     */
+    public static String getSmallFileBase64(String path) {
         return Base64.getEncoder().encodeToString(readFile(path));
     }
 
@@ -406,9 +554,10 @@ public class FileUtils {
      *
      * @param from 源文件
      * @param to   目标文件
+     * @return boolean status
      */
-    public static void moveFile(String from, String to) {
-        new File(from).renameTo(new File(to));
+    public static boolean moveFile(String from, String to) {
+        return new File(from).renameTo(new File(to));
     }
 
     /**
@@ -416,10 +565,11 @@ public class FileUtils {
      *
      * @param path 文件路径
      * @param name 新文件名
+     * @return boolean 状态
      */
-    public static void renameFile(String path, String name) {
+    public static boolean renameFile(String path, String name) {
         File file = new File(path);
-        file.renameTo(new File(file.getParent() + "/" + name));
+        return file.renameTo(new File(file.getParent() + "/" + name));
     }
 
     /**
@@ -430,7 +580,7 @@ public class FileUtils {
      * @return boolean 返回布尔值
      */
     public static boolean checkFileMd5(String path, String md5) {
-        return getFileMd5(path, true).equals(md5.toUpperCase());
+        return Objects.equals(getFileMd5(path, true), md5.toUpperCase());
     }
 
     /**
@@ -441,7 +591,7 @@ public class FileUtils {
      * @return boolean 返回布尔值
      */
     public static boolean checkFileSha1(String path, String sha1) {
-        return getFileSha1(path, true).equals(sha1.toUpperCase());
+        return Objects.equals(getFileSha1(path, true), sha1.toUpperCase());
     }
 
     /**
@@ -463,7 +613,7 @@ public class FileUtils {
      * @return boolean 返回布尔值
      */
     public static boolean checkFileSha256(String path, String sha256) {
-        return getFileSha256(path, true).equals(sha256.toUpperCase());
+        return Objects.equals(getFileSha256(path, true), sha256.toUpperCase());
     }
